Statistics Advance -II
Q1: What are the Probability Mass Function (PMF) and Probability Density Function (PDF)? Explain with
an example.
Q2: What is Cumulative Density Function (CDF)? Explain with an example. Why CDF is used?
Q3: What are some examples of situations where the normal distribution might be used as a model?
Explain how the parameters of the normal distribution relate to the shape of the distribution.
Q4: Explain the importance of Normal Distribution. Give a few real-life examples of Normal
Distribution.
Q5: What is Bernaulli Distribution? Give an Example. What is the difference between Bernoulli
Distribution and Binomial Distribution?
Q6. Consider a dataset with a mean of 50 and a standard deviation of 10. If we assume that the dataset
is normally distributed, what is the probability that a randomly selected observation will be greater
than 60? Use the appropriate formula and show your calculations.
Q7: Explain uniform Distribution with an example.
Q8: What is the z score? State the importance of the z score.
Q9: What is Central Limit Theorem? State the significance of the Central Limit Theorem.
Q10: State the assumptions of the Central Limit Theorem.